litebc_wbc/requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
asttokens==2.0.8
backcall==0.2.0
black==22.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
detectron2==0.6
entrypoints==0.4
executing==1.0.0
fairscale==0.4.12
filelock==3.8.0
focal-loss-torch==0.1.0
fonttools==4.37.1
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.1
hydra-core==1.2.0
idna==3.4
imagecodecs==2022.8.8
imageio==2.21.2
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.9
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.5
networkx==2.8.7
numpy==1.23.3
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.6.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==5.1
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
stack-data==0.5.0
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.11
tomli==2.0.1
torch==1.12.1
torchaudio==0.13.0+cu116
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xgboost==1.6.2
yacs==0.1.8
zipp==3.10.0
